requirements.txt
pandas== 2.1.3
openai==1.3.2
streamlit==1.28.2
python-dotenv==1.0.0